LPS-69343 Split long DataFactory class into several *DataFactory
