+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/DataFactoryConstants.java
@@ -1,0 +1,123 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.tools.sample.sql.builder;
+
+/**
+ * @author Dante Wang
+ */
+public interface DataFactoryConstants {
+
+	public static final String ACCOUNT_LEGAL_NAME = "Liferay, Inc.";
+
+	public static final String ACCOUNT_NAME = "Liferay";
+
+	public static final String ASSET_CATEGORY_NAME_PREFIX = "TestCategory_";
+
+	public static final String ASSET_TAG_NAME_PREFIX = "TestTag_";
+
+	public static final String ASSET_VOCABULARY_NAME_PREFIX = "TestVocabulary_";
+
+	public static final String BLOG_CONTENT_PREFIX = "This is test blog ";
+
+	public static final String BLOG_ENTRY_SUBTITLE_PREFIX =
+		"Subtitle of Test Blog ";
+
+	public static final String BLOG_ENTRY_TITLE_PREFIX = "Test Blog ";
+
+	public static final String BLOG_LAYOUT_NAME = "blogs";
+
+	public static final String BLOG_URL_TITLE_PREFIX = "testblog";
+
+	public static final String COMMERCE_LAYOUT_NAME = "commerce_product";
+
+	public static final String COMPANY_WEBID = "liferay.com";
+
+	public static final String DEFAULT_ASSET_PUBLISHER_PREFERENCE =
+		"default_asset_publisher_preference.xml";
+
+	public static final String DL_DDM_STRUCTURE_CONTENT =
+		"ddm_structure_basic_document.json";
+
+	public static final String DL_DDM_STRUCTURE_LAYOUT_CONTENT =
+		"ddm_structure_layout_basic_document.json";
+
+	public static final String DL_ENTRY_NAME_PREFIX = "TestFile";
+
+	public static final String DL_EXTENSION = "txt";
+
+	public static final String DL_FOLDER_NAME_PREFIX = "Test Folder ";
+
+	public static final String DL_LAYOUT_NAME = "document_library";
+
+	public static final String EMAIL_POSTFIX = "@liferay.com";
+
+	public static final String FIRST_NAME_LIST = "first_names.txt";
+
+	public static final String FORUMS_LAYOUT_NAME = "forums";
+
+	public static final String GREETING_PREFIX = "Welcome ";
+
+	public static final String GROUP_NAME_PREFIX = "Site ";
+
+	public static final String HELLO_WORLD_LAYOUT_NAME = "welcome";
+
+	public static final String JOURNAL_ARTICLE_TITLE_PREFIX =
+		"TestJournalArticle_";
+
+	public static final String JOURNAL_CONTENT_PORTLET_ID =
+		"com_liferay_journal_content_web_portlet_JournalContentPortlet";
+
+	public static final String JOURNAL_DDM_STRUCTURE_CONTENT =
+		"ddm_structure_basic_web_content.json";
+
+	public static final String JOURNAL_DDM_STRUCTURE_LAYOUT_CONTENT =
+		"ddm_structure_layout_basic_web_content.json";
+
+	public static final String JOURNAL_STRUCTURE_KEY = "BASIC-WEB-CONTENT";
+
+	public static final String LANGUAGE_ID = "en_US";
+
+	public static final String LAST_NAME_LIST = "last_names.txt";
+
+	public static final String LAYOUT_COLOR_THEME_ID = "01";
+
+	public static final String LAYOUT_THEME_ID = "classic_WAR_classictheme";
+
+	public static final String MB_BODY_PREFIX = "This is test message ";
+
+	public static final String MB_CATEGORY_NAME_PREFIX = "Test Category ";
+
+	public static final String MB_COMMENT_PREFIX = "This is test comment ";
+
+	public static final String MB_SUBJECT_PREFIX = "Test Message ";
+
+	public static final String REMINDER_QUERY_QUESTION =
+		"What is your screen name?";
+
+	public static final String SAMPLE_USER_NAME = "Sample";
+
+	public static final String USER_PASSWORD = "test";
+
+	public static final String WIKI_LAYOUT_NAME = "wiki";
+
+	public static final String WIKI_NODE_NAME_PREFIX = "Test Node ";
+
+	public static final String WIKI_PAGE_CONTENT_PREFIX = "This is test page ";
+
+	public static final String WIKI_PAGE_FORMAT = "creole";
+
+	public static final String WIKI_PAGE_TITLE_PREFIX = "Test Page ";
+
+}
