+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/SampleSQLBuilder.java
@@ -43,7 +43,6 @@ import java.io.Writer;
 import java.nio.channels.FileChannel;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -67,9 +66,11 @@ public class SampleSQLBuilder {
 
 			properties.load(reader);
 
-			DataFactory dataFactory = new DataFactory(properties);
+			InitContext initContext = new InitContext(properties);
 
-			new SampleSQLBuilder(properties, dataFactory);
+			DataFactory dataFactory = new DataFactory(initContext);
+
+			new SampleSQLBuilder(properties, dataFactory, initContext);
 		}
 		catch (Exception e) {
 			e.printStackTrace();
@@ -86,7 +87,9 @@ public class SampleSQLBuilder {
 		}
 	}
 
-	public SampleSQLBuilder(Properties properties, DataFactory dataFactory)
+	public SampleSQLBuilder(
+			Properties properties, DataFactory dataFactory,
+			InitContext initContext)
 		throws Exception {
 
 		_dbType = DBType.valueOf(
@@ -99,6 +102,7 @@ public class SampleSQLBuilder {
 		_script = properties.getProperty("sample.sql.script");
 
 		_dataFactory = dataFactory;
+		_initContext = initContext;
 
 		// Generic
 
@@ -306,17 +310,19 @@ public class SampleSQLBuilder {
 						createUnsyncBufferedWriter(charPipe.getWriter()),
 						createFileWriter(new File(_outputDir, "sample.sql")));
 
-					FreeMarkerUtil.process(
-						_script,
-						Collections.singletonMap("dataFactory", _dataFactory),
-						sampleSQLWriter);
+					Map<String, Object> context =
+						_dataFactory.getDataFactories();
+
+					context.put("initContext", _initContext);
+
+					FreeMarkerUtil.process(_script, context, sampleSQLWriter);
 				}
 				catch (Throwable t) {
 					_freeMarkerThrowable = t;
 				}
 				finally {
 					try {
-						_dataFactory.closeCSVWriters();
+						_initContext.closeCSVWriters();
 					}
 					catch (IOException ioe) {
 						ioe.printStackTrace();
@@ -409,6 +415,7 @@ public class SampleSQLBuilder {
 	private final DataFactory _dataFactory;
 	private final DBType _dbType;
 	private volatile Throwable _freeMarkerThrowable;
+	private final InitContext _initContext;
 	private final int _optimizeBufferSize;
 	private final String _outputDir;
 	private final String _script;
