+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/SocialActivityDataFactory.java
@@ -1,0 +1,159 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.tools.sample.sql.builder;
+
+import com.liferay.blogs.model.BlogsEntry;
+import com.liferay.blogs.model.BlogsEntryModel;
+import com.liferay.blogs.social.BlogsActivityKeys;
+import com.liferay.document.library.kernel.model.DLFileEntry;
+import com.liferay.document.library.kernel.model.DLFileEntryModel;
+import com.liferay.journal.constants.JournalActivityKeys;
+import com.liferay.journal.model.JournalArticle;
+import com.liferay.journal.model.JournalArticleConstants;
+import com.liferay.journal.model.JournalArticleModel;
+import com.liferay.message.boards.model.MBMessage;
+import com.liferay.message.boards.model.MBMessageModel;
+import com.liferay.message.boards.social.MBActivityKeys;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.util.StringBundler;
+import com.liferay.portlet.documentlibrary.social.DLActivityKeys;
+import com.liferay.portlet.social.model.impl.SocialActivityModelImpl;
+import com.liferay.social.kernel.model.SocialActivityConstants;
+import com.liferay.social.kernel.model.SocialActivityModel;
+import com.liferay.util.SimpleCounter;
+import com.liferay.wiki.model.WikiPage;
+import com.liferay.wiki.social.WikiActivityKeys;
+
+/**
+ * @author Lily Chi
+ */
+public class SocialActivityDataFactory extends BaseDataFactory {
+
+	public SocialActivityDataFactory(InitContext initContext) {
+		super(initContext);
+	}
+
+	public SimpleCounter getSocialActivityCounter() {
+		return _socialActivityCounter;
+	}
+
+	public SocialActivityModel newSocialActivityModel(
+		BlogsEntryModel blogsEntryModel) {
+
+		return _newSocialActivityModel(
+			blogsEntryModel.getGroupId(),
+			getClassNameId(BlogsEntry.class, initContext.getClassNameModels()),
+			blogsEntryModel.getEntryId(), BlogsActivityKeys.ADD_ENTRY,
+			"{\"title\":\"" + blogsEntryModel.getTitle() + "\"}");
+	}
+
+	public SocialActivityModel newSocialActivityModel(
+		DLFileEntryModel dlFileEntryModel) {
+
+		return _newSocialActivityModel(
+			dlFileEntryModel.getGroupId(),
+			getClassNameId(DLFileEntry.class, initContext.getClassNameModels()),
+			dlFileEntryModel.getFileEntryId(), DLActivityKeys.ADD_FILE_ENTRY,
+			StringPool.BLANK);
+	}
+
+	public SocialActivityModel newSocialActivityModel(
+		JournalArticleModel journalArticleModel) {
+
+		int type = JournalActivityKeys.UPDATE_ARTICLE;
+
+		if (journalArticleModel.getVersion() ==
+				JournalArticleConstants.VERSION_DEFAULT) {
+
+			type = JournalActivityKeys.ADD_ARTICLE;
+		}
+
+		return _newSocialActivityModel(
+			journalArticleModel.getGroupId(),
+			getClassNameId(
+				JournalArticle.class, initContext.getClassNameModels()),
+			journalArticleModel.getResourcePrimKey(), type,
+			"{\"title\":\"" + journalArticleModel.getUrlTitle() + "\"}");
+	}
+
+	public SocialActivityModel newSocialActivityModel(
+		MBMessageModel mbMessageModel) {
+
+		long classNameId = mbMessageModel.getClassNameId();
+		long classPK = mbMessageModel.getClassPK();
+
+		int type = 0;
+		String extraData = null;
+
+		if (classNameId == getClassNameId(
+				WikiPage.class, initContext.getClassNameModels())) {
+
+			extraData = "{\"version\":1}";
+
+			type = WikiActivityKeys.ADD_PAGE;
+		}
+		else if (classNameId == 0) {
+			extraData = "{\"title\":\"" + mbMessageModel.getSubject() + "\"}";
+
+			type = MBActivityKeys.ADD_MESSAGE;
+
+			classNameId = getClassNameId(
+				MBMessage.class, initContext.getClassNameModels());
+
+			classPK = mbMessageModel.getMessageId();
+		}
+		else {
+			StringBundler sb = new StringBundler(5);
+
+			sb.append("{\"messageId\": \"");
+			sb.append(mbMessageModel.getMessageId());
+			sb.append("\", \"title\": ");
+			sb.append(mbMessageModel.getSubject());
+			sb.append("}");
+
+			extraData = sb.toString();
+
+			type = SocialActivityConstants.TYPE_ADD_COMMENT;
+		}
+
+		return _newSocialActivityModel(
+			mbMessageModel.getGroupId(), classNameId, classPK, type, extraData);
+	}
+
+	private SocialActivityModel _newSocialActivityModel(
+		long groupId, long classNameId, long classPK, int type,
+		String extraData) {
+
+		SocialActivityModel socialActivityModel = new SocialActivityModelImpl();
+
+		socialActivityModel.setActivityId(_socialActivityCounter.get());
+		socialActivityModel.setGroupId(groupId);
+		socialActivityModel.setCompanyId(initContext.getCompanyId());
+		socialActivityModel.setUserId(initContext.getSampleUserId());
+		socialActivityModel.setCreateDate(_CURRENT_TIME + _timeCounter.get());
+		socialActivityModel.setClassNameId(classNameId);
+		socialActivityModel.setClassPK(classPK);
+		socialActivityModel.setType(type);
+		socialActivityModel.setExtraData(extraData);
+
+		return socialActivityModel;
+	}
+
+	private static final long _CURRENT_TIME = System.currentTimeMillis();
+
+	private final SimpleCounter _socialActivityCounter = new SimpleCounter();
+	private final SimpleCounter _timeCounter = new SimpleCounter();
+
+}
